[feat] #103 drop shadow 함수 modifier extension으로 빼고 주석 추가
